[#71] Anexo em pdf maiores que 5Mb dão tela de erro
Ajusta para fazer o parse da propriedade usando Prop#getLong ao invés de
Prop#getInt

+++ sigaex/src/main/java/br/gov/jfrj/siga/vraptor/ExMultipartConfig.java
@@ -11,10 +11,11 @@ import br.gov.jfrj.siga.base.Prop;
 public class ExMultipartConfig extends DefaultMultipartConfig {  
   
     public long getSizeLimit() {  
-        return Prop.getInt("pdf.tamanho.maximo");
+        return Prop.getLong("pdf.tamanho.maximo");
     }
     
     public long getFileSizeLimit() {
-        return Prop.getInt("pdf.tamanho.maximo");
+        return Prop.getLong("pdf.tamanho.maximo");
     }
-}  
+
+}
